More Linked Lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -10,12 +10,12 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int num = 0;
+	size_t num = 0;
 
 	while (h)
 	{
 		num++;
-		printf("%d\n", (h->n));
+		printf("%lu\n", (h->n));
 		h = h->next;
 	}
 	return (num);
